+++ tests/ConfigurationTest.php
@@ -72,4 +72,24 @@ class ConfigurationTest extends TestCase
         Configuration::setCreditKey('something-secret');
         $this->assertEquals('something-secret', Configuration::getCreditKey());
     }
+
+    public function testGetReportsApiUrl()
+    {
+        $this->assertEquals('https://portal.flitt.com', Configuration::getReportsApiUrl());
+        Configuration::setReportsApiUrl('portal.example.com');
+        $this->assertEquals('https://portal.example.com', Configuration::getReportsApiUrl());
+        Configuration::setReportsApiUrl('portal.flitt.com');
+    }
+
+    public function testSetReportsApplicationId()
+    {
+        Configuration::setReportsApplicationId('1234');
+        $this->assertEquals('1234', Configuration::getReportsApplicationId());
+    }
+
+    public function testSetReportsApplicationKey()
+    {
+        Configuration::setReportsApplicationKey('something-secret');
+        $this->assertEquals('something-secret', Configuration::getReportsApplicationKey());
+    }
 }